Dirty tracking is now disabled when initializing.
The way that tracking was disabled while serializing records skipped the
after initialize hook, which can break things that are expecting it to
fire when serialzing a remote call response.

Since create always sends all of the records, we don't really need dirty
tracking when initializing because all it tells us is that something was
once nil.