Reinstate Model#add_errors_from_response
When Model#add_errors_from_response was deprecated, checking if the response
actually defines errors was removed as well. Since not all responses
have define errors, this causes exceptions.

The problem with Model#add_errors_from_response wasn't that it took a
response, but that it relied on @last_response. Instead of deprecating
the whole method, just deprecate that behavior.