Adds setters for has_one and belongs_to
We weren't defining setters during these associations before. Next steps will
add a setter for has_many. I feel like that's a little different because it deals
with a number of records rather than just one, so there needs to be some check that
the current method does not do.