Reset attributes before assigning response attributes
Why? The response proto will only return attributes explicitly set by
the server. This is efficient. What we want now is an efficient way to
reset our internal state. With this new reset before assignment, we will
always honor the response sent by the server. Previously, we were not
doing this.

Suppose the server always ignores writing an attribute, but the client
attempts to update this attribute anyway:

```
some_model.update(:the_server_ignores_this => true)
```

The correct behavior here should be to reset the
`:the_server_ignores_this` attribue back to `false` but that's not what
happens. The attribute will stay `true` because the server did not
explicitly set the value to `false`, so the attribute was missing from
`response_proto.to_hash`.

But, with the new reset, we'll always refresh the model's state to
reflect the server. Always.

NOTE: Active record does not refresh internal state after an update
statement, but this is a different problem. An update statement does not
return the new result set to the client. Active record would need to
perform another lookup to refresh the internal state, and that's slow.
But active remote has the latest state. The server just gave it to us,
so we should us it.