Add support for registering types
In preparation for transitioning to Active Model 5.2, add support for
registering types that can be used to define attributes without using
the existing :type or :typecaster options:

  attribute :name, :string

The :type and :typecaster options still work, but this brings the
interface for the attribute method inline with what it is in Active
Model 5.2.

Additionally, update the Protobuf serializer to use the new type
registry rather than maintaining a map of typecasters directly.