Make query attributes more permissive
Active Record considers numeric values present only when `zero?` is
false. This is incompatible with Protobuf enums since they are usually
zero-based, but will also cause numeric values not to be sent over the
wire when they are 0, 0.0, etc. This may be fine for AR, but doesn't
work so well here in Active Remote.